+++ changes.md
@@ -4,3 +4,6 @@ Release/revision numbering is major only, no minors - i.e. 1,2,3... not 1.1, 1.2
 
 #### 2017-11-27 R6
 * Baseline release to GitHub (from private repo).
+
+#### 2018-01-22 R8
+* Changed sscanf calls to use SCNu8 macro for uint_8t vars, SCNd8 macro for int_8t vars